+++ mtg/src/test/scala/com/github/gchudnov/mtg/internal/rel/MeetsSpec.scala
@@ -1,4 +1,4 @@
-package com.github.gchudnov.mtg.internal
+package com.github.gchudnov.mtg.internal.rel
 
 import com.github.gchudnov.mtg.Arbitraries.*
 import com.github.gchudnov.mtg.Interval
@@ -6,7 +6,6 @@ import com.github.gchudnov.mtg.Mark
 import com.github.gchudnov.mtg.Domain
 import com.github.gchudnov.mtg.TestSpec
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks.*
-import com.github.gchudnov.mtg.internal.rel.IntervalRelAssert
 
 /**
  * Meets, IsMetBy
